Added contact to home page, adding copyright to end of it
