+++ my-portfolio/src/App.css
@@ -16,72 +16,6 @@
 }
 
 /*----------------------------------------------------------------------------------------*/
-/* Contact Page */
-.contact-page {
-  width: 100%;
-  height: 100%;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.contact-info {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  margin-bottom: 32px;
-}
-
-.contact-info a {
-  border: 2px #18134C solid;
-  padding: 5px;
-  border-radius: 20px;
-  color: #18134C;
-  width: auto;
-}
-
-.contact-info a span {
-  color: #18134C;
-}
-
-.contact-label {
-  color: #18134C;
-  font-weight: 600;
-}
-
-.contact-area {
-  background-color: #00D7FF;
-}
-
-.placeholder-Image {
-  width: 100%;
-
-  display: flex;
-  justify-content: center;
-  margin-bottom: 32px;
-}
-
-.form-manage {
-  border: 2px #18134C solid;
-  border-radius: 20px;
-  padding: 10px;
-}
-
-@media screen and (max-width: 768px) {
-  .contact-info {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 5px;
-  }
-
-  .contact-info p {
-    width: auto;
-  }  
-}
-
-/*----------------------------------------------------------------------------------------*/
 /* Responsive */
 @media screen and (max-width: 769px) {
   html {
