+++ my-portfolio/src/pages/1-start-page/start-page.css
@@ -91,34 +91,6 @@
   border-radius: 15px;
 }
 
-/* .copyright-drake {
-  gap: 10px;
-  color: #18134C;
-  width: 100%;
-  height: fit-content;  
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: row;
-}
-
-.copyright-drake-1 {
-  position: absolute;
-  bottom: 0;
-  color: #72FFFF;
-  width: 100%;
-  height: fit-content;  
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: row;
-  gap: 10px;
-} */
-
-/* .svg-icon {
-  color: #0096FF;
-} */
-
 .start-cv-btn {
   margin-top: 1.5rem;
   padding: 0.7em 2.5em;
@@ -156,16 +128,6 @@
     display: flex;
     text-align: start;
   }
-
-
-  /* .copyright-drake {
-    left: 20%;
-    color: #18134C;
-    
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-  } */
 }
 
 @media screen and (max-width: 1100px) {
