+++ my-portfolio/src/pages/2-about-page/about-page.js
@@ -68,10 +68,6 @@ const AboutPage = () => {
                     </div>
                 </div> */}
             </div>
-
-            {/* <div className="page-body" id="contact-block">
-                <ContactMe />
-            </div> */}
         </>
     );
 }
