+++ my-portfolio/src/pages/6-project-page/project-techStack-list.js
@@ -3,7 +3,7 @@ const ProjectStack = ({techStacks}) => {
         <>
             <div className="tech-stack-list">
                 {techStacks.map((tech, index) => (
-                    <a key={index} className='tech-stack-element'>{tech}</a>
+                    <p key={index} className='tech-stack-element'>{tech}</p>
                 ))}
             </div>
         </>
