+++ my-portfolio/src/pages/7-contact-page/contact-page.css
@@ -1,0 +1,147 @@
+
+.page-body-contact {
+    /* height: 90vh; */
+    /* height: 100%; */
+    height: max-content;
+    width: 100%;
+  
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding: 10vh 4% 2% 4%;
+}
+
+/*----------------------------------------------------------------------------------------*/
+/* Contact Page */
+.contact-page {
+    width: 100%;
+    height: 100%;
+  
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    }
+  
+.contact-form-frame {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+.contact-info {
+    display: flex;
+    flex-direction: row;
+    gap: 30%;
+    justify-content: center;
+}
+  
+.contact-info-detail {
+    border: 2px var(--text-color) solid;
+    padding: 10px;
+    border-radius: 20px;
+    color: var(--text-color);
+    width: max-content;
+
+    display: flex;
+    flex-direction: row;
+    gap: 10px;
+    align-items: center;
+    justify-content: center;
+
+    margin: 0;
+}
+  
+.contact-info-detail span {
+    color: var(--text-color-highlight);
+}
+  
+.contact-label {
+    color: var(--text-color);
+    font-weight: 600;
+}
+  
+.contact-area {
+    background-color: var(--button-color);
+    color: var(--button-text-color);
+}
+  
+  .placeholder-Image {
+    width: 100%;
+  
+    display: flex;
+    justify-content: center;
+    margin-bottom: 32px;
+  }
+  
+  .form-manage {
+    border: 2px var(--text-color) solid;
+    border-radius: 20px;
+    padding: 10px;
+  }
+
+  .start-cv-btn-contact {
+    margin-top: 1.5rem;
+    padding: 0.7em 2.5em;
+    border-radius: 25px;
+    font-size: 1rem;
+    font-weight: 400;
+    border: 0;
+  
+    background-color: var(--background-color);
+    color: var(--text-color-highlight);
+    border: 2px var(--text-color-highlight) solid;
+    cursor: pointer;
+  }
+  
+  .start-cv-btn-contact:hover {
+    /* background-color: #72FFFF; */
+    background-color: var(--button-color-hover);
+    color: var(--button-text-color-hover);
+    box-shadow: 0 0 10px #72FFFF;
+    border: 2px var(--button-text-color-hover) solid;
+    transition: all .3s ease-in-out;
+    font-weight: 600;
+    transform: scale(1.25); 
+  }
+
+  .contact-links {
+    color: var(--text-color);
+  }
+
+  .copyright-drake {
+    gap: 10px;
+    color: var(--text-color);
+    width: 100%;
+    height: fit-content;  
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: row;
+  }
+
+  .copyright-drake h5 {
+    margin: 0;
+  }
+  
+  @media screen and (max-width: 768px) {
+    .contact-info {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      gap: 5px;
+    }
+  
+    .contact-info p {
+      width: auto;
+    }  
+
+    .copyright-drake {
+      left: 20%;
+      color: #18134C;
+      
+      display: flex;
+      align-items: center;
+      flex-direction: row;
+    }
+  }
